refactor: move post coordinator dashboard stuff related to its own file
